+++ packages/zephyr/src/Menus/components/MenuItemDescription.tsx
@@ -1,6 +1,8 @@
-import React from 'react';
+import React, { memo } from 'react';
 import { Text, TextProps } from '../../Text';
 
-export const MenuItemDescription = ({ tx, ...restOfProps }: TextProps) => (
+export const MenuItemDescription = memo(({ tx, ...restOfProps }: TextProps) => (
   <Text tx={{ mt: 2, color: 'pigeon500', ...(tx as any) }} variant="text-ui-14" {...restOfProps} />
-);
+));
+
+MenuItemDescription.displayName = 'MenuItemDescription';
