+++ packages/zephyr/src/RadixMenu/components/MenuItemDivider.tsx
@@ -1,11 +1,7 @@
-import React, { memo } from 'react';
+import React, { memo, FC, ComponentProps } from 'react';
 import { Separator } from '@radix-ui/react-dropdown-menu';
-import styled from 'styled-components';
 
-import { BoxProps } from '../../Box';
-import { inlineStyles } from '../utils/inlineStyles';
-
-const StyledSeperator = styled(Separator)<{ tx?: BoxProps['tx'] }>(inlineStyles);
+import { Box } from '../../Box';
 
 export interface MenuDividerProps {
   isTop?: boolean;
@@ -13,7 +9,8 @@ export interface MenuDividerProps {
 
 export const MenuItemDivider = memo(({ isTop }: MenuDividerProps) => {
   return (
-    <StyledSeperator
+    <Box
+      as={Separator as FC<Omit<ComponentProps<typeof Separator>, 'as' | 'ref'>>}
       tx={{ backgroundColor: 'pigeon100', height: 1, border: 0, mt: isTop ? 0 : 8, mb: 8 }}
     />
   );
