--- packages/zephyr/src/RadixMenu/components/MenuItemLabel.tsx
@@ -1,19 +1,0 @@
-import React, { memo } from 'react';
-
-import { Text, TextProps } from '../../Text';
-
-export const MenuItemLabel = memo(({ tx, ...rest }: TextProps) => (
-  <Text
-    tx={{
-      color: 'inherit',
-      whiteSpace: 'nowrap',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      ...(tx as any),
-    }}
-    variant="text-ui-16"
-    {...rest}
-  />
-));
-
-MenuItemLabel.displayName = 'MenuItemLabel';
