+++ packages/zephyr/src/RadixMenu/index.tsx
@@ -1,0 +1,3 @@
+export * from './abstractions/DropdownMenu';
+export * from './components/MenuItemDivider';
+export * from './components/MenuItem';
