--- packages/zephyr/src/RadixMenu/utils/inlineStyles.ts
@@ -1,3 +1,0 @@
-import css from '@styled-system/css';
-
-export const inlineStyles = ({ tx, theme }: any) => css({ ...tx })(theme);
